move edit and add bottom sheet to shared components.
