+++ app/src/main/java/com/amsterdam/cutetudee/presentation/component/AddOrEditTaskBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.presentation.screens.component
+package com.amsterdam.cutetudee.presentation.component
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.background
@@ -26,13 +26,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.amsterdam.cutetudee.R
 import com.amsterdam.cutetudee.domain.entity.Task
-import com.amsterdam.cutetudee.presentation.component.CustomBottomSheet
-import com.amsterdam.cutetudee.presentation.component.CustomDatePickerDialog
-import com.amsterdam.cutetudee.presentation.component.CustomTextField
-import com.amsterdam.cutetudee.presentation.component.GradientFilledButton
-import com.amsterdam.cutetudee.presentation.component.OutlineButton
-import com.amsterdam.cutetudee.presentation.component.ReadOnlyCustomTextField
-import com.amsterdam.cutetudee.presentation.component.SelectedBadgedCategory
 import com.amsterdam.cutetudee.presentation.component.chip.priority.PriorityChip
 import com.amsterdam.cutetudee.presentation.component.chip.priority.PriorityUi
 import com.amsterdam.cutetudee.presentation.screens.common.AddEditTaskInteractionListener
